Merge beta10 into master

+++ library/prng/erand48.c
@@ -1,5 +1,5 @@
 /*
- * $Id: stdlib_erand48.c,v 1.0 2022-08-06 10:36:26 clib2devs Exp $
+ * $Id: stdlib_erand48.c,v 1.0 2022-08-06 10:36:26 clib4devs Exp $
 */
 #ifndef _STDLIB_HEADERS_H
 #include "../stdlib/stdlib_headers.h"
@@ -13,5 +13,5 @@ erand48(unsigned short s[3]) {
         uint64_t u;
         double f;
     } x = {0x3ff0000000000000ULL | __rand48_step(s, __seed48 + 3) << 4};
-    return x.f - 1.0;
+    return x.f - (double) 1.0;
 }
