+++ library/profile/profil.c
@@ -1,5 +1,5 @@
 /*
- * $Id: profile_profil.c,v 1.0 2021-01-21 10:08:32 clib2devs Exp $
+ * $Id: profile_profil.c,v 1.0 2021-01-21 10:08:32 clib4devs Exp $
 */
 
 #include <proto/exec.h>
@@ -34,8 +34,8 @@ GetCounterStart(void) {
             TAG_DONE);
 
     /* Timebase ticks at 1/4 of FSB */
-    bit0time = 8.0 / (double) fsb;
-    count = (uint32)(0.01 / bit0time);
+    bit0time = (double) 8.0 / (double) fsb;
+    count = (uint32)((double) 0.01 / bit0time);
 
     return 0x80000000 - count;
 }
